Added downloading and importing plugin to build.sh script. Using UNITY_ADS define to avoid compiler error when project hasn't been imported

+++ build.sh
@@ -1,69 +1,88 @@
-# ./build.sh platform unity-version
-# e.g.
-#   ./build.sh android 5.4.0f1
-#   ./build.sh ios 5.4.0f1
+# NOTE: This is a test script used by Unity Ads team to verify Ads SDK versions
+# against different Unity versions. Parameters and usage of this script may change
+# in future versions, but is included here as example of how to automate this
+# process, which might be relevant to others
 
-if [ -z "$1" ]; then
-    echo "Build Unity Ads SDK example project"
+if [ -z "$2" ]; then
+    echo "Build Unity Ads SDK example project on OS X"
     echo
-    echo "./build.sh platform unity-version"
+    echo "./build.sh platform unity-path [Ads SDK version]"
     echo
-    echo "Example:"
-    echo "  ./build.sh android 5.4.0f1"
-    echo "  ./build.sh ios 5.4.0f1"
+    echo "Examples:"
+    echo "  ./build.sh android \"/Application/Unity 5.4.0f3\""
+    echo "  ./build.sh ios \"/Application/Unity 5.4.0f3\" 2.0.3"
     echo
-    echo "Specified Unity version e.g. '5.4.0f1' must be installed on the machine, in folder /Applications/Unity (version)"
+    echo "Unity version e.g. \"5.4.0f3\" must be installed on the machine in specified folder"
+    echo "If specifying SDK version (2.0.0 or higher), this will be downloaded and imported into the project automatically"
     exit 1
 fi
 
-# echo Downloading and importing package from $3...
-# curl -s -o UnityAds.unitypackage $3 
-# rc=$?; if [[ $rc != 0 ]]; then
-#     echo "Failed to download package"
-#     exit $rc
-# fi
-
-# Import package into project
-
-UNITY="/Applications/Unity $2/Unity.app/Contents/MacOS/Unity"
+UNITY="$2/Unity.app/Contents/MacOS/Unity"
 if [ ! -f "$UNITY" ]; then
-    echo "Could not find Unity installed at '$UNITY'. Please verify it's installed and available in that location"
+    echo "Could not find Unity executable in '$UNITY'. Please verify it's installed and available in that location"
     exit 1
 fi
 
-if [ -f $(pwd)/UnityAds.unitypackage ]; then
-    echo "Import Unity Ads asset store package..."
+if [ ! -z "$3" ]; then
+    SDK_URL="http://cdn.unityads.unity3d.com/unitypackage/$3/UnityAds.unitypackage"
+    echo Downloading Ads SDK from $SDK_URL...
+    curl -s -o UnityAds.unitypackage $SDK_URL
+    rc=$?; if [[ $rc != 0 ]]; then
+        echo "Failed to download package"
+        exit $rc
+    fi
+
+    # Import package into project
+    echo "Importing Ads SDK plugin..."
     "$UNITY" -projectPath "$(pwd)" -importPackage "$(pwd)/UnityAds.unitypackage" -batchMode -quit
     rc=$?; if [[ $rc != 0 ]]; then
         echo "Importing package failed. Please check ~/Library/Logs/Unity/Editor.log"
         exit $rc
     fi
+
+    # UNITY_ADS define is used to be able to import plugin from command line without compile erros in project
+    echo "Setting UNITY_ADS define..."
+    "$UNITY" -projectPath "$(pwd)" -executeMethod AutoBuilder.EnableAds -batchMode -quit
+    rc=$?; if [[ $rc != 0 ]]; then
+        echo "Defining UNITY_ADS failed. Please check ~/Library/Logs/Unity/Editor.log"
+        exit $rc
+    fi
 fi
 
 if [ "$1" == "android" ]; then
-    echo Building and running project with Unity $2 \(Android\)...
-    rm Builds/Android.apk
+    echo Building project for Android...
+    APK_PATH="Builds/Android.apk"
+    if [ -f $APK_PATH ]; then
+        rm $APK_PATH
+    fi
     "$UNITY" -projectPath $(pwd) -executeMethod AutoBuilder.PerformAndroidBuild -batchMode -quit
     rc=$?; if [[ $rc != 0 ]]; then
         echo "Unity build for Android failed. Please check ~/Library/Logs/Unity/Editor.log"
         exit $rc
     fi
 
-    echo Installing on Android device...
-    adb install -r Builds/Android.apk
+    if [ ! -f $APK_PATH ]; then
+        echo "Failed to build APK file ($APK_PATH)"
+        exit 1
+    fi
 
-    adb shell am start -S -a android.intent.action.MAIN -n com.unity3d.UnityAdsAssetStoreTest/com.unity3d.player.UnityPlayerActivity
+    # TODO: Enable automatic deployment on device
+    # echo Installing on Android device...
+    # adb install -r Builds/Android.apk
+    # adb shell am start -S -a android.intent.action.MAIN -n com.unity3d.UnityAdsAssetStoreTest/com.unity3d.player.UnityPlayerActivity
 fi
 
 if [ "$1" == "ios" ]; then
-    echo Building project with Unity $2 \(iOS\)...
-    rm -rf Builds/iOS
+    echo Building project for iOS...
+    if [ -d Builds/iOS ]; then
+        rm -rf Builds/iOS
+    fi
     "$UNITY" -projectPath $(pwd) -executeMethod AutoBuilder.PerformiOSBuild -batchMode -quit
     rc=$?; if [[ $rc != 0 ]]; then
         echo "Unity build for iOS failed. Please check ~/Library/Logs/Unity/Editor.log"
         exit $rc
     fi
 
-    open Builds/iOS/Unity-iPhone.xcodeproj
+    # open Builds/iOS/Unity-iPhone.xcodeproj
     # xcodebuild -project Builds/iOS/Unity-iPhone.xcodeproj
 fi
